% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupGammaGateR.R
\name{collateGroupGammaGateR}
\alias{collateGroupGammaGateR}
\title{Binds groupGammaGateR}
\usage{
collateGroupGammaGateR(...)
}
\arguments{
\item{...}{groupGammaGateR objects.}
}
\value{
Analogous to GammaGateR, groupGammaGateR simpky returns lists in the name of slides, each list being corresponding GammaGateR object.
}
\description{
Extends the c() for groupGammaGateR object
}
\details{
Fits cfGMM models to each marker channel in a matrix of marker channels for multiple slides
}
