% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag.R
\name{diag_contrast}
\alias{diag_contrast}
\title{Diagnostic plot - contrast}
\usage{
diag_contrast(
  fit1,
  fit2,
  marker = 1,
  component = 2,
  title = "Contrastive Diagnostic Plot",
  fit.names = c("fit1", "fit2"),
  boundaries = NULL,
  ...
)
}
\arguments{
\item{fit1}{A groupGammaGateR object.}

\item{fit2}{Another groupGammaGateR object, possibly that is a refit of fit1 with updated constraints.}

\item{marker}{Marker to be plotted, defaults to the first one.}

\item{component}{Integer specifying which component to plot, 1 is unexpressed nonzero cells, 2 is expressed cells.}

\item{title}{Title of plot.}

\item{fit.names}{Names of fit objects. Default to c("fit1", "fit2")}

\item{boundaries}{Add vertical boundary lines to plot.}

\item{...}{Arguments passed to plot.GammaGateR}
}
\description{
This can be seen as an overlay of the scatter plot diagnostic plot. While side-by-side comparison of the plots
might have scale problem, "overlay" the scatter plot resolves this issue. The corresponding slides are also connected
with dashed line, making the changes easier to spot.
}
\details{
Plot a histogram of a GammaGateR model. Plots one model.
Takes a GammaGateR object and plots the histogram with the fitted model and parameter values.
}
