% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histograms.R
\name{evaluateGroupGammaGateR}
\alias{evaluateGroupGammaGateR}
\title{Compare binary classification result using Cohen's Kappa or Adjusted Rand Index}
\usage{
evaluateGroupGammaGateR(
  ...,
  standard,
  batch,
  method = c("cohen", "rand"),
  plot.type = c("boxplot", "scatter")
)
}
\arguments{
\item{...}{Matrices to compare to standard labels. Rows are cells, columns are markers.}

\item{standard}{A matrix indicating gold/silver standard positive cells.}

\item{batch}{Split kappa computation by this variable. Similar to slide_id.}

\item{method}{"cohen" the default, calculates cohen's kappa; "rand" calculates adjusted rand index.}

\item{plot.type}{"boxplot" the default; recommend "scatter" when number of slides is small.}
}
\description{
This function can compare the agreement of classification result against a gold/silver standard.
The result can be returned as in either a data frame of Cohen's Kappa or a side-by-side box plot.
}
\details{
Plot a histogram of a GammaGateR model. Plots one model.
Takes a GammaGateR object and plots the histogram with the fitted model and parameter values.
}
