% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GammaGateR.R
\name{GammaGateR}
\alias{GammaGateR}
\title{Fits a gamma mixture model to aid clustering of mIF imaging data}
\usage{
GammaGateR(
  expressionMarkers,
  boundaryMarkers = NULL,
  qboundaryMarkers = NULL,
  subBatch = NULL,
  ...
)
}
\arguments{
\item{expressionMarkers}{An ncell by nmarker matrix of expression values. Assumed to be coarsely normalized and transformed.}

\item{boundaryMarkers}{A nmarker list of 4x4 matrices giving the boundaries for the modes of the unexpressed and expressed cell distributions.}

\item{qboundaryMarkers}{A nmarker list of 4x4 matrices giving the qauntile boundaries for the modes for the unexpressed and expressed cell distributions.}

\item{subBatch}{If there are multiple subBatch on a slide, subBatch can be used to return probability estimates independently for each region.}

\item{...}{Arguments passed to cfGMM function}
}
\value{
\item{expressionZ}{The monotone proportional density of second component}
\item{expressionX}{The input expression value}
\item{expressionW}{The raw proportional density of second component}
\item{params}{The monotone proportional density of second component}
\item{fit}{cfGMM::cfGMM fitting output}
\item{subBatch}{same as the subBatch in the argument}
}
\description{
Takes a cfGMM model and forces the right-most component
probabilities to be monotonic increasing with respect
to the input values.
}
\details{
Fits cfGMM models to each marker channel in a matrix of marker channels for one slide
}
