% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GammaGateR.R
\name{GammaGateRX}
\alias{GammaGateRX}
\title{Fits a gamma mixture model to a vector of expression data}
\usage{
GammaGateRX(x, constraints = NULL, subBatch = NULL, nn0 = 200, ...)
}
\arguments{
\item{x}{A vector of expression values. Assumed to be coarsely normalized and transformed.}

\item{constraints}{A nmarker list of 4x4 matrices giving the boundaries for the modes of the unexpressed and expressed cell distributions.}

\item{subBatch}{If there are multiple subBatch on a slide, the vector subBatch can be used to return probability estimates independently for each region.}

\item{nn0}{If there are fewer than nn0, then it will not fit the cfGMM. Default is 200.}

\item{...}{Arguments passed to cfGMM function}
}
\description{
Fits a gamma mixture model to a vector of expression data
}
\details{
Takes a cfGMM model and forces the right-most component
probabilities to be monotonic increasing with respect
to the input values.
}
