% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histograms.R
\name{get.mode}
\alias{get.mode}
\title{Get mode from parameters}
\usage{
get.mode(par.mat)
}
\arguments{
\item{par.mat}{parameter matrix from GammaGateR or groupGammaGateR object.}
}
\description{
Get mode from parameters
}
\details{
Get mode from a parameter matrix
}
