% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupGammaGateR.R
\name{groupGammaGateR}
\alias{groupGammaGateR}
\title{Fits a gamma mixture model to mIF imaging data with multiple slides at one time}
\usage{
groupGammaGateR(
  expressionMarkers,
  slide,
  boundaryMarkers = NULL,
  qboundaryMarkers = NULL,
  subBatch = NULL,
  n.cores = 1,
  ...
)
}
\arguments{
\item{expressionMarkers}{An ncell by nmarker matrix of expression values. Assumed to be coarsely normalized and transformed.}

\item{slide}{Single vector of length n. Identifies slide number for each cell.}

\item{boundaryMarkers}{A nmarker list of 4x4 matrices giving the boundaries for the modes of the unexpressed and expressed cell distributions.}

\item{qboundaryMarkers}{A nmarker list of 4x4 matrices giving the qauntile boundaries for the modes for the unexpressed and expressed cell distributions.}

\item{subBatch}{If there are multiple subBatch on a slide, subBatch can be used to return probability estimates independently for each region.}

\item{n.cores}{Number of cores. Default to one core.}

\item{...}{Arguments passed to GammaGateR.}
}
\value{
Analogous to GammaGateR, groupGammaGateR simpky returns lists in the name of slides, each list being corresponding GammaGateR object.
}
\description{
Takes a cfGMM model and forces the right-most component
probabilities to be monotonic increasing with respect
to the input values. Apply this model to multiple slides with multiple channels.
}
\details{
Fits cfGMM models to each marker channel in a matrix of marker channels for multiple slides
}
