% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histograms.R
\name{hist_sr_constrast}
\alias{hist_sr_constrast}
\title{Contrastive density curve on histogram}
\usage{
hist_sr_constrast(
  fit1,
  fit2,
  marker = 1,
  slide = 1,
  subBatch = 1,
  title = NULL,
  ...
)
}
\arguments{
\item{fit1}{A GammaGateR or groupGammaGateR object.}

\item{fit2}{Another GammaGateR or groupGammaGateR object, possibly that is a refit of fit1 with updated constraints. Has to be of the same class as fit1.}

\item{marker}{Marker to be plotted, defaults to the first one.}

\item{slide}{Slide to be plotted, defaults to the first one.}

\item{subBatch}{subBatch to be plotted, defaults to the first one.}

\item{title}{Title of plot.}

\item{...}{Arguments passed to plot.GammaGateR}
}
\description{
This function can compare two fitted model of a marker on a slide, from GammaGateR/groupGammaGateR object.
}
\details{
Plot histogram constrast of two GammaGateR or groupGammaGateR model. Plots one model.
Takes a GammaGateR object and plots the histogram with the fitted model and parameter values.
}
