% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histograms.R
\name{hist_sr_constrast_inner}
\alias{hist_sr_constrast_inner}
\title{Histogram contrast (inner function, not shown)}
\usage{
hist_sr_constrast_inner(
  fit1,
  fit2,
  marker = 1,
  subBatch = 1,
  title = NULL,
  ...
)
}
\arguments{
\item{fit1}{A GammaGateR object.}

\item{fit2}{Another GammaGateR object, possibly that is a refit of fit1 with updated constraints.}

\item{marker}{Marker to be plotted, defaults to the first one.}

\item{subBatch}{subBatch to be plotted, defaults to the first one.}

\item{title}{Title of plot.}

\item{...}{Arguments passed to plot.GammaGateR}
}
\description{
Histogram contrast (inner function, not shown)
}
\details{
Plot a histogram of a GammaGateR model. Plots one model.
Takes a GammaGateR object and plots the histogram with the fitted model and parameter values.
}
