% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monoZ.R
\name{monoZ}
\alias{monoZ}
\title{Make cfGMM probabilities monotonic}
\usage{
monoZ(fit)
}
\arguments{
\item{fit}{A cfGMM model object}
}
\description{
Takes a cfGMM model and forces the right-most component
probabilities to be monotonic increasing with respect
to the input values.
}
