% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histograms.R
\name{plot.GammaGateR}
\alias{plot.GammaGateR}
\title{Histogram of data with density curve}
\usage{
\method{plot}{GammaGateR}(
  x,
  marker = 1,
  subBatch = 1,
  zero_include = FALSE,
  breaks = 40,
  title = "histogram of x",
  boundary = NULL,
  hist = TRUE,
  dens = TRUE,
  tabl = FALSE,
  col = "forestgreen",
  p = NULL,
  print = TRUE,
  ...
)
}
\arguments{
\item{x}{A GammaGateR object.}

\item{marker}{Marker to be plotted, defaults to the first one.}

\item{subBatch}{subBatch to be plotted, defaults to the first one.}

\item{zero_include}{Logical whether the histogram should include zeros. Default is FALSE.}

\item{breaks}{number of breaks for histogram}

\item{title}{Title of plot.}

\item{boundary}{Vertical line drawn on plot to indicate boundary constraint.}

\item{hist}{Logical indicating whether to plot histogram and base layer of plot.}

\item{dens}{Logical indicating whether to draw parametric density over histogram.}

\item{tabl}{Logical indicating whether to add table of parameter values.}

\item{col}{Color of density.}

\item{p}{ggplot object to add layers too. If hist=FALSE this must be specified.}

\item{print}{logical whether to print the plot.}

\item{...}{Arguments passed to cfGMM function}
}
\value{
This function gives a plot of histogram and optional fitted curve from a GammaGateR object.
}
\description{
This plot funtion can plot the histogram of the data and with the addition of density
curve (optional) overlay on top the histogram.
}
\details{
Plot a histogram of a GammaGateR model. Plots one model.
Takes a GammaGateR object and plots the histogram with the fitted model and parameter values.
}
